pass Observable subscribe on RestaurantViewModel extract util methods
