+++ app/src/main/java/com/picone/go4lunch/presentation/utils/RecyclerViewItemClickUtil.java
@@ -1,4 +1,4 @@
-package com.picone.go4lunch.presentation.ui.utils;
+package com.picone.go4lunch.presentation.utils;
 
 import android.view.View;
