+++ app/src/main/java/com/picone/go4lunch/presentation/viewModels/RestaurantViewModel.java
@@ -10,6 +10,7 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 import com.picone.core.domain.entity.Restaurant;
+import com.picone.core.domain.entity.RestaurantPOJO.RestaurantPOJO;
 import com.picone.core.domain.entity.User;
 import com.picone.core.domain.entity.UserDailySchedule;
 import com.picone.core.domain.entity.predictionPOJO.Prediction;
@@ -37,10 +38,13 @@ import java.util.List;
 import java.util.Locale;
 
 import io.reactivex.Observable;
+import io.reactivex.ObservableOnSubscribe;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 
 import static com.picone.go4lunch.presentation.ui.fragment.MapsFragment.MAPS_KEY;
+import static com.picone.go4lunch.presentation.utils.CreateRestaurantFromMapsUtil.createRestaurant;
+import static com.picone.go4lunch.presentation.utils.CreateRestaurantFromMapsUtil.formatOpeningHours;
 
 public class RestaurantViewModel extends ViewModel {
 
@@ -129,24 +133,64 @@ public class RestaurantViewModel extends ViewModel {
 
     @SuppressWarnings("ResultOfMethodCallIgnored")
     @SuppressLint("CheckResult")
-    public void getRestaurantFromMaps() {
+    public void getRestaurantFromMaps_() {
         Location mCurrentLocation = locationMutableLiveData.getValue();
         locationMutableLiveData.setValue(mCurrentLocation);
-        fetchRestaurantFromPlaceInteractor.fetchRestaurantFromPlace(mCurrentLocation, MAPS_KEY)
+        List<Restaurant> restaurantsFromMap = new ArrayList<>();
+        fetchRestaurantFromPlaceInteractor.fetchRestaurantFromPlace_(mCurrentLocation, MAPS_KEY)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .flatMap(allRestaurants -> {
-                    Log.i("TAG", "getRestaurantFromMaps: " + allRestaurants);
-                    for (Restaurant restaurant : allRestaurants) {
-                        fetchRestaurantDetailFromPlaceInteractor.getRestaurantDetail(restaurant, MAPS_KEY);
-                        fetchRestaurantDistanceInteractor.getRestaurantDistance(restaurant, mCurrentLocation, MAPS_KEY);
-                    }
-                    return updateAllRestaurantsWithPersistedValues(allRestaurants);
-                })
-                .subscribe(allRestaurants -> {
+                .flatMap(nearBySearch -> Observable.create((ObservableOnSubscribe<List<Restaurant>>)
+                        emitter -> {
+                            if (nearBySearch.getStatus().equals("OK")) {
+                                for (RestaurantPOJO restaurantPOJO : nearBySearch.getRestaurantPOJOS()) {
+                                    Restaurant restaurant = createRestaurant(restaurantPOJO, MAPS_KEY);
+                                    if (!restaurantsFromMap.contains(restaurant))
+                                        restaurantsFromMap.add(restaurant);
+                                }
+                            }
+                            Log.i("TAG", "getRestaurantFromMaps_: nearBySearch" + restaurantsFromMap.get(0).getName());
+                            emitter.onNext(restaurantsFromMap);
+                        }))
+                .flatMap(this::fetchPlaceDetail)
+                .flatMap(restaurants -> fetchPlaceDistance(mCurrentLocation, restaurants))
+                .subscribe(restaurants -> {
+                    Log.i("TAG", "getRestaurantFromMaps_: subscribe" + restaurants.get(0).getDistance());
                     if (filteredRestaurantMutableLveData.getValue() == null || allRestaurantsMutableLiveData.getValue() == null)
-                        allRestaurantsMutableLiveData.setValue(allRestaurants);
+                        allRestaurantsMutableLiveData.setValue(restaurants);
                 });
+    }
+
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    private Observable<List<Restaurant>> fetchPlaceDistance(Location mCurrentLocation, List<Restaurant> restaurants) {
+        return Observable.create(emitter -> {
+            for (Restaurant restaurant : restaurants)
+                fetchRestaurantDistanceInteractor.getRestaurantDistance_(restaurant, mCurrentLocation, MAPS_KEY)
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(restaurantDistance -> {
+                            restaurant.setDistance(restaurantDistance.getRows().get(0).getElements().get(0).getDistance().getText());
+                            Log.i("TAG", "getRestaurantFromMaps_: restaurant distance" + restaurant.getDistance());
+                        });
+            emitter.onNext(restaurants);
+        });
+    }
+
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    private Observable<List<Restaurant>> fetchPlaceDetail(List<Restaurant> restaurants) {
+        return Observable.create(emitter -> {
+            for (Restaurant restaurant : restaurants)
+                fetchRestaurantDetailFromPlaceInteractor.getRestaurantDetail_(restaurant, MAPS_KEY)
+                        .subscribeOn(Schedulers.io())
+                        .observeOn(AndroidSchedulers.mainThread())
+                        .subscribe(restaurantDetail -> {
+                            restaurant.setPhoneNumber(restaurantDetail.getResult().getFormattedPhoneNumber());
+                            restaurant.setWebsite(restaurantDetail.getResult().getWebsite());
+                            restaurant.setOpeningHours(formatOpeningHours(restaurantDetail));
+                            Log.i("TAG", "getRestaurantFromMaps_: restaurant detail" + restaurant.getPhoneNumber());
+                        });
+            emitter.onNext(restaurants);
+        });
     }
 
     @SuppressWarnings("ResultOfMethodCallIgnored")
